updated docs and comments
