+++ Sources/PureduxStore/Observer.swift
@@ -1,6 +1,6 @@
 //
 //  File.swift
-//  
+//
 //
 //  Created by Sergey Kazakov on 01.12.2020.
 //
@@ -13,21 +13,21 @@ public enum ObserverStatus {
     case dead
 }
 
-/**
- Observer can be subscribed to Store to handle state updates
-*/
+
+///Observer can be subscribed to Store to handle state updates
+///
 public struct Observer<State>: Hashable {
-    /**
-     Closure that handles  Observer's status
-     - Parameter  status: observer status to handle
-    */
+
+    ///     Closure that handles  Observer's status
+    ///     - Parameter  status: observer status to handle
+    ///
     public typealias StatusHandler = (_ status: ObserverStatus) -> Void
 
-    /**
-     Observer's main closure that handle State changes and calls complete handler
-        - Parameter state: new observed State
-        - Parameter complete: complete handler that Observer calls when the work is done
-    */
+
+    ///     Observer's main closure that handle State changes and calls complete handler
+    ///        - Parameter state: new observed State
+    ///        - Parameter complete: complete handler that Observer calls when the work is done
+    ///
     public typealias StateHandler = (_ state: State, _ complete: @escaping  StatusHandler) -> Void
 
     let id: UUID
@@ -45,16 +45,16 @@ public struct Observer<State>: Hashable {
 }
 
 public extension Observer {
-    /**
-     Initializes a new store Observer
 
-     - Parameter observe:Is a closure that receive state and StatusHandler as parameters.
-     When Observer decides to unsubscribe from store it passes .dead status parameter to status handler.
-
-     - Returns: Observer
-
-     Observer handles new states, sent by the Store.
-     */
+    ///     Initializes a new store Observer
+    ///
+    ///     - Parameter observe:Is a closure that receive state and StatusHandler as parameters.
+    ///     When Observer decides to unsubscribe from store it passes .dead status parameter to status handler.
+    ///
+    ///     - Returns: Observer
+    ///
+    ///     Observer handles new states, sent by the Store.
+    ///
     init(observe: @escaping StateHandler) {
         id = UUID()
         self.observeClosure = observe
